dev-requirements.txt
# ! Arquivo deve ser travado: não deve ser alterado por pessoas estudantes

-e .
-r requirements.txt

black==23.3.0
click==8.1.3
coverage==7.2.7
factory-boy==3.2.1
Faker==18.9.0
flake8==6.0.0
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
mypy==1.3.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest-cov==4.1.0
pytest-dependency@git+https://github.com/betrybe/pytest-dependency@984f9d7d083870d091e8862a9b9c33fdf815b8d9
pytest-json==0.4.0
pytest==7.3.1
python-dateutil==2.8.2
rich==13.4.2
shellingham==1.5.0.post1
six==1.16.0
typer==0.9.0
typing_extensions==4.6.2
wheel==0.40.0
